requirements.txt
notion-client==2.2.1
python-dotenv==1.0.0
requests==2.31.0
keyring==25.2.1
pytest==8.3.2
reportlab==4.2.0